+++ monks-tokenbar-api.js
@@ -1,7 +1,7 @@
 import { MonksTokenBar, log, i18n, setting, MTB_MOVEMENT_TYPE } from "./monks-tokenbar.js";
 import { AssignXPApp } from "./apps/assignxp.js";
-import { SavingThrowApp } from "./apps/savingthrow.js";
-import { ContestedRollApp } from "./apps/contestedroll.js";
+import { SavingThrowApp, SavingThrow } from "./apps/savingthrow.js";
+import { ContestedRollApp, ContestedRoll } from "./apps/contestedroll.js";
 import { LootablesApp } from "./apps/lootables.js";
 
 export class MonksTokenBarAPI {