+++ systems/base-rolls.js
@@ -10,12 +10,13 @@ export class BaseRolls {
             id: "dice", text: "Dice", cssclass: "dice-group", groups: { "1d2": "1d2", "1d4": "1d4", "1d6": "1d6", "1d8": "1d8", "1d10": "1d10", "1d12": "1d12", "1d20": "1d20", "1d100": "1d100" }
         }];
 
+        /*
         this._defaultSetting = {
             stat1 : "attributes.ac.value",
             stat2 : "",
             icon1 : "fa-shield-alt",
             icon2 : "fa-eye"
-        }
+        }*/
     }
 
     get _supportedSystem() {
@@ -39,6 +40,10 @@ export class BaseRolls {
 
     get showXP() {
         return true;
+    }
+
+    get defaultStats() {
+        return [];
     }
 
     getButtons() {
@@ -117,6 +122,10 @@ export class BaseRolls {
 
     defaultContested() {
         return null;
+    }
+
+    dynamicRequest(tokens) {
+        return [];
     }
 
     roll({ id }, callback, e) {